copy/paste && navbar
Added the basic layout skeleton for the pages corresponding with what is on the dashboard. This will be adjusted to be different when the styling is applied

Started on the styling for the navbar. Wanted to started with the news part with trying to make a icon button for the nav links, using news coo birds for a template first