been awhile
it has been awhile. I have been finishing this wedding planning. My lord, I am a bit overwhelmed.. Well, that's life. I want to make this site for my in home display within a mirror. Want to really do this project. I am getting tired of unfinished projects that could and/or are a good idea.. That's it, I will code again in the morning.